+++ app/controllers/settings/users_controller.rb
@@ -6,7 +6,7 @@ class Settings::UsersController < ApplicationController
   before_action :authenticate_admin!, except: %i[export import]
 
   def index
-    @users = User.order(created_at: :desc)
+    @users = User.non_deleted.order(created_at: :desc)
   end
 
   def edit
@@ -40,11 +40,18 @@ class Settings::UsersController < ApplicationController
   def destroy
     @user = User.find(params[:id])
 
-    if @user.destroy
-      redirect_to settings_users_url, notice: 'User was successfully deleted.'
-    else
-      redirect_to settings_users_url, notice: 'User could not be deleted.', status: :unprocessable_content
+    unless @user.can_delete_account?
+      redirect_to settings_users_url,
+                  alert: 'Cannot delete account while being owner of a family which has other members.',
+                  status: :unprocessable_content
+      return
     end
+
+    @user.mark_as_deleted! unless @user.deleted?
+    Users::DestroyJob.perform_later(@user.id)
+
+    redirect_to settings_users_url,
+                notice: 'User deletion has been initiated. The account will be fully removed shortly.'
   end
 
   def export