+++ app/jobs/area_visits_calculating_job.rb
@@ -7,7 +7,8 @@ class AreaVisitsCalculatingJob < ApplicationJob
   sidekiq_options retry: false
 
   def perform(user_id)
-    user = User.find(user_id)
+    user = find_non_deleted_user(user_id)
+    return unless user
 
     with_user_timezone(user) do
       areas = user.areas