+++ app/jobs/import/photoprism_geodata_job.rb
@@ -5,7 +5,8 @@ class Import::PhotoprismGeodataJob < ApplicationJob
   sidekiq_options retry: false
 
   def perform(user_id)
-    user = User.find(user_id)
+    user = find_non_deleted_user(user_id)
+    return unless user
 
     Photoprism::ImportGeodata.new(user).call
   end