+++ app/jobs/place_visits_calculating_job.rb
@@ -5,7 +5,9 @@ class PlaceVisitsCalculatingJob < ApplicationJob
   sidekiq_options retry: false
 
   def perform(user_id)
-    user = User.find(user_id)
+    user = find_non_deleted_user(user_id)
+    return unless user
+
     places = user.places # Only user-owned places (with user_id)
 
     Places::Visits::Create.new(user, places).call